User response, entity

+++ src/main/java/com/donks/bci/bcichallenge/service/impl/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.donks.bci.bcichallenge.service.impl;
 
 import com.donks.bci.bcichallenge.entity.UserEntity;
+import com.donks.bci.bcichallenge.exceptions.NotFoundException;
+import com.donks.bci.bcichallenge.exceptions.NotValidException;
 import com.donks.bci.bcichallenge.repository.UserRepository;
 import com.donks.bci.bcichallenge.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -24,16 +26,16 @@ public class UserServiceImpl implements UserService {
     private Pattern emailRegex;
 
     public UserEntity createUser(UserEntity user) {
+//        throw new NotValidException("TESTING");
         if (user.getId() != null && userRepository.existsById(user.getId()))
-            throw new RuntimeException("User already exists");
+            throw new NotValidException("User already exists");
 
-        if (user.getEmail() == null)
-            throw new RuntimeException("Email vacio");
+        if (user.getEmail() == null) throw new NotValidException("Empty email");
 
-        if (!emailRegex.matcher(user.getEmail()).matches())
-            throw new RuntimeException("Email no valido");
+        if (!emailRegex.matcher(user.getEmail()).matches()) throw new NotValidException("Invalid email");
 
-
+        if (userRepository.findByEmail(user.getEmail()).isPresent())
+            throw new NotValidException("Email already in use");
 
         return userRepository.save(user);
 
@@ -41,7 +43,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Optional<UserEntity> findById(UUID id) {
+        if (!userRepository.existsById(id)) throw new NotFoundException("User not found");
+
         return userRepository.findById(id);
+
     }
 
     @Override
@@ -50,6 +55,8 @@ public class UserServiceImpl implements UserService {
     }
 
     public boolean deleteById(UUID id) {
+        if (!userRepository.existsById(id)) throw new NotFoundException("User not found");
+
         userRepository.deleteById(id);
         return !userRepository.existsById(id);
     }
